:recycle: user can authenticate their wax account w/discogs as a new or existing user
