+++ app/serializers/user_serializer.rb
@@ -1,3 +1,3 @@
 class UserSerializer < ActiveModel::Serializer
-  attributes :name, :email, :discogs_id, :oauth_token, :oauth_token_secret, :wax_username
+  attributes :name, :email, :discogs_id, :oauth_token, :oauth_token_secret, :wax_username, :id
 end
