+++ config.ru
@@ -2,4 +2,16 @@
 
 require_relative 'config/environment'
 
+# require 'rack/cors'
+# use Rack::Cors do
+
+#  # allow all origins in development
+#  allow do
+#    origins '*'
+#    resource '*',
+#        :headers => :any,
+#        :methods => [:get, :post, :delete, :put, :options]
+#  end
+# end
+
 run Rails.application
