+++ config/initializers/cors.rb
